When loading the IDL code in C or Perl: generate the array_positions based on the sequence in which the fields appear in the IDL file (rather than from attributes in the <field> elements).
In most cases this change will make no difference, because the
fields appear in sequence.  However for classes "brn" and "mvr" the 
sequences of fields in the IDL are scrambled.  For those classes, 
this change will effectively renumber the fields.

This commit does not affect the corresponding Python or Java code.
So don't go editing the IDL yet to remove the array_position
attributes.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@13015 dcc99617-32d9-48b4-a31d-7c20da2025e4