Add a serial search method that returns an array of MFHD hashes for a given bib record
This enables libraries to display multiple MFHD records per bib record, for example if
a library has serial holdings for a given journal in microfiche, microfilm, and print,
the holdings can be represented separately.

Note that we are not currently identifying which MFHD record is providing the
holdings coverage; the likely short-term plan will be to display location information
such as 852 $c.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@13224 dcc99617-32d9-48b4-a31d-7c20da2025e4