Patch from Jeff Godin (MLC/TADL) which:
Fixes the mixed content warnings from IE7 and IE8 "out of the box" without
breaking sites that use static servers for CSS and Javascript.

This sets OILS_PROTOCOL based on the environment variable HTTPS (which
Apache sets to "on" for https:// requests).

Thus, a page loaded via https:// gets https:// urls, and a page loaded via
http:// gets http:// urls.

To accommodate sites with static servers that can not support https,
there's a new OILS_OPAC_STATIC_PROTOCOL variable in eg_vhost.conf
which can be set to "http" to force JS/CSS to be loaded over http://
(you'll get the same warnings that you got before).



git-svn-id: svn://svn.open-ils.org/ILS/trunk@13288 dcc99617-32d9-48b4-a31d-7c20da2025e4