More tweaks to oils_cstore.c:
1. In osrfAppInitialize(): reuse the growing_buffer method_name
instead of repeatedly destroying and recreating it.  That should
save hundreds of mallocs and frees in the initialization.

2. Also in osrfAppInitialize(): correct an inaccurate debug
message, and rearrange a few lines for clarity.

3. Throughout: replace the dbi_get_* functions, which do lookups
by name, with the corresponding dbi_get_*_idx functions, which
do lookups by index.  Lookups by index are faster than lookups
by name, because they don't have to do a search.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@13508 dcc99617-32d9-48b4-a31d-7c20da2025e4