Patch from Warren Layton for support of RIS format in SuperCat - thanks Warren!
===

I've put together another SuperCat feed, this type for the RIS format.
This can be used with most bibliographic citation software, such as
Reference Manager and Endnote. Combined with the previous "MARC text"
feed for RefWorks, most software should be covered.

I have only included the changes to add the feed and the XSLT file.
The patch (against trunk) doesn't include any changes to the OPAC, but
I'm willing to put something together if needed.

The MARC21slim2RIS.xsl transform makes a few assumptions:
* Any authors listed in 100, 110, or 111 fields are considered
"primary authors" (A1)
* Any individuals listed in 700 fields are considered primary authors
(A1) if their last name is found in the statement of responsibility
(245c). Otherwise, they are considered secondary authors (A2). It's
not 100% accurate, but it's certainly a step up from just turning all
700-level authors into A2s.
* Corporate authors listed in 710 fields are always considered
secondary authors (A2).
* Meetings in 711 fields are always converted into "Title Series" (T3)
(i.e., a talk at a conference would have the conference name listed as
the "Title Series", with the title of the talk listed as "T1", primary
title).
* Apart from field 520 (used for Abstracts, "N2"), no 5xx fields are used
* Subfields in the 650 and 651 fields are used as keywords ("KW")

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
   have the right to submit it under the open source license
   indicated in the file; or

(b) The contribution is based upon previous work that, to the best
   of my knowledge, is covered under an appropriate open source
   license and I have the right under that license to submit that
   work with modifications, whether created in whole or in part
   by me, under the same open source license (unless I am
   permitted to submit under a different license), as indicated
   in the file; or

(c) The contribution was provided directly to me by some other
   person who certified (a), (b) or (c) and I have not modified
   it.

(d) I understand and agree that this project and the contribution
   are public and that a record of the contribution (including all
   personal information I submit with it, including my sign-off) is
   maintained indefinitely and may be redistributed consistent with
   this project or the open source license(s) involved.

Signed-off-by: Warren Layton


git-svn-id: svn://svn.open-ils.org/ILS/trunk@13700 dcc99617-32d9-48b4-a31d-7c20da2025e4