Patch from Galen Charlton to add Version to application.ini
==
The attached patch restores the Version key to the staff client's application.ini.  This key is now required; without it, xulrunner-bin --install-app on OS X will silently crash, leading to an incomplete and non-functional installation of the client on that platform.

Note that when the client is built, the value of Version is replaced by the build ID.





git-svn-id: svn://svn.open-ils.org/ILS/trunk@13803 dcc99617-32d9-48b4-a31d-7c20da2025e4