Fix a horrible regression when renewing multiple items at once in Items Out.
Referencing loop variables in Javascript with closures is dangerous, so one strategy is to do something like this:

funcs = [];
for (var i = 0; i < my_array.length; i++ ) {

	/* Bad */
	// funcs.push( function(){ do_something( my_array[i] ); } );
	/* Better */
	funcs.push( function(safe_value){ return function(){ do_something( safe_value ); } }( my_array[i] ) );
}

In our case, our generated function accidentally referenced a value dependent on the loop variable instead of the corresponding argument of the function generator.

So we had multiple async renewal calls that depending on the timing, could try to renew the same item.  To further add insult to injury, this could potentially put the database in an inconsistent 
state wtih duplicate circulations.



git-svn-id: svn://svn.open-ils.org/ILS/trunk@13889 dcc99617-32d9-48b4-a31d-7c20da2025e4