Update script patch from Joe Atzberger.
===

The purpose of this script is to consolidate a lot of the annoying and error-prone tasks associated with an upgrade for a developer, including make, make install, the xulrunner client built and autogen.

Considerations:
* Run as user "opensrf"
* opensrf needs sudo (again, targeting developers, not production) for apache stop/start, chown and make install
* Assumes opensrf has two SVN (or git-svn) repos: OpenILS and OpenSRF
* Both repos should be already configured (as in ./configure)

Detailed usage notes available via -h option.  Both repo directories and the install directory can be specified on the command line.  Try it with -t (test mode) to see feedback without making any changes.

===

TODO.  Add option to use brick_ctl.sh instead of osrf_ctl.sh for service start/stop




git-svn-id: svn://svn.open-ils.org/ILS/trunk@15015 dcc99617-32d9-48b4-a31d-7c20da2025e4