DB update patch/script from Joe Atzberger.
===
Attachment one is a database update script based on work Bill sent to the list.  The previous mechanism was to call psql -f once per file, but that also required the psql password to be entered once per file and didn't check to see if the call failed to try again or at least bail out on subsequent files (that might require the earlier ones to succeed.  Given that typing (or even pasting) 20+ consecutive blind password entires is error-prone, I replaced that with a pipe to single psql call. 

Pending acceptance, the foreseeable goal will be to integrate this script with the update.sh script and have a single call to apply both code and data changes from an updated repo. 
===



git-svn-id: svn://svn.open-ils.org/ILS/trunk@15016 dcc99617-32d9-48b4-a31d-7c20da2025e4