Hack: create symbolic links with names demanded by Translate Toolkit
GNU gettext recommended layout is foo.dtd/ll.po, where "ll" represents
the locale of the translated resource.

But Translate Toolkit doesn't offer any command line flags to tell it
what kind of file it's supposed to generate; it relies on the filename:

The file extension of the original Mozilla file is required to tell the
Toolkit how to do the conversion. Therefore, a file like foo.dtd must be
named foo.dtd.po in order to po2moz to recognise it as a DTD file.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@15074 dcc99617-32d9-48b4-a31d-7c20da2025e4