Add Table of Contents capability to OpenLibrary added content plugin.
Hard code the URLs for OpenLibrary in the module, as requiring the URLs
to be set in opensrf.xml is an unnecessary configuration step.

Add a GPL header to the module, crediting David Christensen as the originator
and myself as the extender.

Some other miscellaneous code cleanup.

TODO: Enable toc_json() method; probably by refactoring toc_html().


git-svn-id: svn://svn.open-ils.org/ILS/trunk@15089 dcc99617-32d9-48b4-a31d-7c20da2025e4