1. Renamed getSourceDefinition() to getRelation(), since "SourceDefinition" could be taken to refer to a subquery defined in the IDL.
2. In getRelation(): when returning a source_definition instead of a table name
or view name, build it in a single allocated buffer instead of a growing_buffer,
in order to save a malloc() and free().

3. In osrfAppChildInit(): Use and reuse a single growing_buffer to build the
SELECT statements, instead of allocating and freeing one repeatedly.

4. In osrfAppChildInit(): if getRelation() returns NULL, skip the class and go
on to the next one, since any attempt to SELECT from it is doomed anyway.

5. In osrfAppChildInit(): eliminate several unnecessary casts.

M    Open-ILS/src/c-apps/oils_cstore.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@15651 dcc99617-32d9-48b4-a31d-7c20da2025e4