Tweaked oilsAuthVerifyPassword(), mostly for clarity:
1. Addad a doxygen-style comment block to document it, plus several comments
in the body of the function.

2. A slight rearrangement: get the password from the user object after getting
the seed from memcache, instead of before.  That way if the memcache call
fails we don't need to free the copy of the user object's password.

3. Moved the declaration of ret (the return code) closer to its first use.

4. Plugged a potential (albeit improbable) memory leak in the case of an
early exit.

Also: several minor tweaks to white space and comments elsewhere.

M    Open-ILS/src/c-apps/oils_auth.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@15886 dcc99617-32d9-48b4-a31d-7c20da2025e4