Performance tweak to oilsAuthComplete().
Formerly, if authenticating a barcode rather tnan a user name, we would
look up the same barcode twice in actor.card: first to get the user id
for a lookup in actor.usr, and again to determine whether the card was
active.  Now we use a single lookup for both purposes.  In other respects
the validation logic is unchanged.

The oilsAuthCheckCard() function, which performed the second lookup, has
been eliminated.

The oilsUtilsFetchUserByBarcode() function, in oils_utils.c, is no
longer called from anywhere and may be eliminated.

Also added a doxygen-style comment block to document oilsUtilsFetchUserByBarcode(),
and tweaked few comments in the body of the function.

M    Open-ILS/src/c-apps/oils_auth.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@15916 dcc99617-32d9-48b4-a31d-7c20da2025e4