Acq: more work on the claiming UI. It begins to be workable, but still rough.
Definitely still have to fix at least these things:

Changes to lineitems' claim policies and initiation of claims against LIs
aren't reflected in realtime in the LI table. You still need a reload for
that so far.

The "Consider individual copies for claiming" needs to take you to the
speicific lineitems' claim copies dialog, not just to the PO in general.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16290 dcc99617-32d9-48b4-a31d-7c20da2025e4