Acq: the XULTermLoader dijit now returns the first columns of CSV rows
Could be improved later to allow more control over what data you get back,
but it works fine for the current uses (reading a file of one-term-per-line
for unified search and reading a CSV file where the first column is of
interest in the Bib record->lineitem interface).

Also added a little support to the LI table/pager for remember what lineitems
are selected even after you move on to the next page.  Nothing really takes
full advantage of this yet, though.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16432 dcc99617-32d9-48b4-a31d-7c20da2025e4