Bug fixes in verifyObjectPCRUD():
1. Don't throw an exception just because a foreign key is null.

2. When looking for local or foreign contexts, always do a lookup of
the current row, even if you already have an image of that row in
hand.  The image you have may not include the foreign key(s) you need.

The latter fix is simple but inefficient.  It should be possible to
avoid the extra lookup most of the time, and maybe all of the time,
for the search, retrieve, and id_list methods.  However that will
involve additional complications, not yet implemented.  Let's make
it correct first, and worry about efficiency later.

M    Open-ILS/src/c-apps/oils_sql.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16504 dcc99617-32d9-48b4-a31d-7c20da2025e4