Various tweaks:
1. When resetting the timeout for a session: if the session object
carries an endtime, and the expiration time is within a designated
interval, extend the expiration time a bit instead of reapplying the
entire original timeout from the current time.

This feature is unreachable at the moment because it is designed for
persistent sessions, which aren't otherwise implemented yet.  The
purpose is to avoid rudely terminating an active session, but without
extending it too much.

2. Whenever possible, use integral types instead of doubles.

3. Instead of calling osrfCacheSetExpire(), call osrfCachePutObject()
directly, which does the same thing with less overhead.

4. Tinkered with white space and comments here and there.

M    Open-ILS/src/c-apps/oils_auth.c


git-svn-id: svn://svn.open-ils.org/ILS/trunk@16586 dcc99617-32d9-48b4-a31d-7c20da2025e4