Bugfixes: undef handling and legacy script support setting.
The legacy setting was a major bug.  It was not possible to disable legacy scripts
except by *removing* the setting.  That is, if you set it to 'false',
we failed to interpret that here in SIP to mean FALSE.  Instead we
looked at it as a non-zero-length string and therefore TRUE!

This patch also prevents warnings from unitialized values (undef concatenation), like:
Use of uninitialized value in concatenation (.) or string at /openils/lib/perl5/OpenILS/SIP/Patron.pm line 110.

Added a little formatting and whitespace cleanup to address display along w/ undef handling.

Corrected an error screen message also.

git-svn-id: svn://svn.open-ils.org/ILS/trunk@16600 dcc99617-32d9-48b4-a31d-7c20da2025e4